requirements.txt
beautifulsoup4==4.12.2
ddddocr==1.4.8
Pillow==10.0.0
pkcs7==0.1.2
pycryptodome==3.18.0
PyExecJS==1.5.1
Requests==2.31.0
